+++ student-store-ui/src/components/App/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Navbar from "../Navbar/Navbar";
 import Sidebar from "../Sidebar/Sidebar";
 import Home from "../Home/Home";
-import About from "../About";
+import About from "../About/About";
 import "./App.css";
 import Footer from "../Footer/Footer";
 
@@ -12,15 +12,15 @@ export default function App() {
     <div className="app-container">
       <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/about" element={<About />} />
+        {/* <Route path="/" element={<Home />} /> */}
+        <Route path="/" element={<About />} />
       </Routes>
-        <main className="app-content">
+        {/* <main className="app-content"> */}
           {/* <Navbar /> */}
           {/* <Sidebar /> */}
           {/* <Home /> */}
           {/* <Footer /> */}
-        </main>
+        {/* </main> */}
       </BrowserRouter>
     </div>
   )
